- silenced lots of warnings for deliberate type conversions.
The ones in d_net.cpp are not in active code.

+++ source/core/automap.cpp
@@ -184,7 +184,7 @@ static void CalcMapBounds()
 static void AutomapControl(const DVector2& cangvect)
 {
 	static double nonsharedtimer;
-	double ms = screen->FrameTime;
+	double ms = (double)screen->FrameTime;
 	double interval;
 
 	if (nonsharedtimer > 0 || ms < nonsharedtimer)
