+++ source/core/r_data/gldefs.cpp
@@ -1566,7 +1566,7 @@ class GLDefsParser
 						ExtraUniformCVARData* oldextra = nullptr;
 						sc.MustGetFloat();
 
-						val.Float = oldval.Float = sc.Float;
+						val.Float = oldval.Float = (float)sc.Float;
 
 						if (!Args->CheckParm ("-shaderuniformtest"))
 						{
